pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pythonutilities"
version = "0.1.0"
description = "The Bocan Online Developer Toolkit"
requires-python = ">=3.10"
dependencies = ["typer", "rich", "pyyaml"]

[project.scripts]
bo-add  = "bocan.apps.add.cli:bocan.app"
bo-help = "bocan.apps.help.cli:bocan.app"
bo-list = "bocan.apps.list.cli:bocan.app"
bo-new  = "bocan.apps.new.cli:bocan.app"
bo-open = "bocan.apps.open.cli:bocan.app"

[tool.setuptools]
package-dir = {"" = "source"}

[tool.setuptools.packages.find]
where = ["source"]
include = ["*"]
exclude = ["core"]
namespaces = false
